backend/requirements.txt
flask==2.3.3
google-generativeai==0.3.1
python-dotenv==1.0.0
flask-cors==4.0.0
langchain==0.1.5
langchain-openai==0.0.5
langchain-community==0.0.15
langchain-text-splitters==0.0.1
pypdf==3.17.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
chromadb==0.4.22
weaviate-client==3.25.2
tqdm==4.66.1
pymongo==4.5.0
psycopg2-binary==2.9.9
pymilvus==2.2.12
pgvector==0.2.3
sqlalchemy==2.0.27
jsonrpcserver==5.0.9
jsonrpcclient==4.0.2